Subscene-Foreignonly v1.9.6

+++ service.py
@@ -19,7 +19,6 @@ import re
 import difflib
 from operator import itemgetter
 
-
 __addon__ = xbmcaddon.Addon()
 __author__ = __addon__.getAddonInfo('author')
 __scriptid__ = __addon__.getAddonInfo('id')
@@ -75,11 +74,12 @@ def rmtree(path):
     xbmcvfs.rmdir(path)
 
 
-try:
-    rmtree(__temp__)
-except:
-    pass
-xbmcvfs.mkdirs(__temp__)
+# cleaning up temp directory if it exists
+if xbmcvfs.exists(__temp__):
+    try:
+        rmtree(__temp__)
+    except:
+        pass
 
 
 def find_movie(content, title, year):
@@ -89,6 +89,7 @@ def find_movie(content, title, year):
         import html
     else:
         html = HTMLParser.HTMLParser()
+
     for secmatches in re.finditer(search_section_pattern, content, re.IGNORECASE | re.DOTALL):
         log(__name__, secmatches.group('section'))
         for matches in re.finditer(movie_season_pattern, secmatches.group('content'), re.IGNORECASE | re.DOTALL):
@@ -155,7 +156,7 @@ def find_tv_show_season(content, tvshow, season):
         import html
     else:
         html = HTMLParser.HTMLParser()
-
+        
     for matches in re.finditer(movie_season_pattern, content, re.IGNORECASE | re.DOTALL):
         found_title = matches.group('title')
         found_title = html.unescape(found_title)
@@ -204,7 +205,6 @@ def append_subtitle(item):
     listitem.setProperty("hearing_imp", 'true' if item["hearing_imp"] else 'false')
     #ForeignOnly change
     listitem.setProperty("forced", 'true')
-
 
     # below arguments are optional, it can be used to pass any info needed in download function
     # anything after "action=download&" will be sent to addon once user clicks listed subtitle to downlaod
@@ -409,8 +409,13 @@ def download(link, episode=""):
         tempdir = os.path.join(__temp__, str(uid))
     else:
         tempdir = os.path.join(__temp__, unicode(uid))
-    if not xbmcvfs.mkdirs(tempdir):
+
+    try:
+        os.makedirs(tempdir)
+    except OSError:
         log(__name__, "Failed to create temp directory " + tempdir)
+    else:
+        log(__name__, "Successfully created temp directory " + tempdir)
 
     content, response_url = geturl(link)
     content = str(content)
@@ -488,22 +493,64 @@ def download(link, episode=""):
 
         if packed:
             xbmc.sleep(500)
-            log(__name__, "Extracting '%s' to '%s'" % (local_tmp_file, tempdir))
             if sys.version_info.major == 3:
-                (dirs, files) = xbmcvfs.listdir('%s' % (local_tmp_file))
-                src = os.path.join(local_tmp_file, files[0])
-                dest = os.path.join(tempdir, files[0])
-                log(__name__, 'copy %s to %s' %(src, dest))
-                xbmcvfs.copy(src, dest)
-            else:
-                xbmc.executebuiltin(('XBMC.Extract("%s","%s")' % (local_tmp_file, tempdir,)).encode('utf-8'), True)
+                log(__name__, "Checking '%s' for subtitle files to copy" % (local_tmp_file))
+                if sys.platform == "linux" or sys.platform == "linux2":
+                    log(__name__, "Platform identified as Linux")
+                    #Kodi on linux does not understand 'archive://' protocol
+                    (dirs, files) = xbmcvfs.listdir('%s' % xbmcvfs.translatePath(local_tmp_file))
+                else:
+                    log(__name__, "Platform identified as Non-Linux")
+                    #Kodi on windows and possibly Android requires archive:// protocol, so testing both
+                    (dirs, files) = xbmcvfs.listdir('archive:\\\\%s' % xbmcvfs.translatePath(urllib.parse.quote_plus(local_tmp_file)))
+                    if len(files) == 0:
+                        (dirs, files) = xbmcvfs.listdir('%s' % xbmcvfs.translatePath(local_tmp_file))
+                for file in files:
+                    dest = os.path.join(tempdir, file)
+                    log(__name__, "=== Found subtitle file %s" % dest)
+                    if sys.platform == "linux" or sys.platform == "linux2":
+                        #Kodi on linux does not understand 'archive://' protocol
+                        src = os.path.join(local_tmp_file, file)
+                        log(__name__, "trying to copy '%s' to '%s'" % (src, dest))
+                        if not xbmcvfs.copy(src, dest):
+                            log(__name__, "copying failed")
+                        else:
+                            log(__name__, "copying succeeded")
+                    else:
+                        #Kodi on windows and possibly Android requires archive:// protocol, so testing both
+                        src = xbmcvfs.translatePath(os.path.join("archive:\\\\%s" % urllib.parse.quote_plus(local_tmp_file), file))
+                        log(__name__, "trying to copy '%s' to '%s'" % (src, dest))
+                        if not xbmcvfs.copy(src, dest):
+                            log(__name__, "copying failed")
+                            #trying again
+                            src = os.path.join(local_tmp_file, file)
+                            log(__name__, "trying to copy '%s' to '%s'" % (src, dest))
+                            if not xbmcvfs.copy(src, dest):
+                                log(__name__, "copying failed")
+                            else:
+                                log(__name__, "copying succeeded")
+                        else:
+                            log(__name__, "copying succeeded")
 
+                    subtitle_list.append(dest)
+            else:
+                log(__name__, "Extracting '%s' to '%s'" % (local_tmp_file, tempdir))
+                xbmc.executebuiltin(('XBMC.Extract("%s","%s")' % (local_tmp_file, tempdir,)).encode('utf-8'), True)
+                for file in xbmcvfs.listdir(local_tmp_file)[1]:
+                    file = os.path.join(tempdir, file)
+                    if os.path.splitext(file)[1] in exts:
+                        log(__name__, "=== Found subtitle file %s" % file)
+                        subtitle_list.append(file)
+            
         episode_pattern = None
         if episode != '':
             episode_pattern = re.compile(get_episode_pattern(episode), re.IGNORECASE)
 
+        log(__name__, "Checking temp dir subfolders for subtitle files...")
         for dir in xbmcvfs.listdir(tempdir)[0]:
+            log(__name__, "Check dir subfolder %s" % dir)
             for file in xbmcvfs.listdir(os.path.join(tempdir, dir))[1]:
+                log(__name__, "Check dir subfolder file %s" % file)
                 if os.path.splitext(file)[1] in exts:
                     log(__name__, 'match '+episode+' '+file)
                     if episode_pattern and not episode_pattern.search(file):
@@ -511,7 +558,9 @@ def download(link, episode=""):
                     log(__name__, "=== returning subtitle file %s" % file)
                     subtitle_list.append(os.path.join(tempdir, dir, file))
 
+        log(__name__, "Checking temp dir for subtitle files...")
         for file in xbmcvfs.listdir(tempdir)[1]:
+            log(__name__, "Check dir file %s" % file)
             if os.path.splitext(file)[1] in exts:
                 log(__name__, 'match '+episode+' '+file)
                 if episode_pattern and not episode_pattern.search(file):
@@ -522,13 +571,17 @@ def download(link, episode=""):
         if len(subtitle_list) == 0:
             if sys.version_info.major == 3:
                 if episode:
+                    log(__name__, "=== Could not find matching episode in subtitle pack")
                     xbmc.executebuiltin('Notification(%s,%s)' % (__scriptname__, __language__(32002)))
                 else:
+                    log(__name__, "=== Download didn't contain a subtitle file")
                     xbmc.executebuiltin('Notification(%s,%s)' % (__scriptname__, __language__(32003)))
             else:
                 if episode:
+                    log(__name__, "=== Could not find matching episode in subtitle pack")
                     xbmc.executebuiltin((u'Notification(%s,%s)' % (__scriptname__, __language__(32002))).encode('utf-8'))
                 else:
+                    log(__name__, "=== Download didn't contain a subtitle file")
                     xbmc.executebuiltin((u'Notification(%s,%s)' % (__scriptname__, __language__(32003))).encode('utf-8'))
 
     return subtitle_list
@@ -627,5 +680,6 @@ elif params['action'] == 'download':
         listitem = xbmcgui.ListItem(label=sub)
         xbmcplugin.addDirectoryItem(handle=int(sys.argv[1]), url=sub, listitem=listitem, isFolder=False)
 
+
 xbmcplugin.endOfDirectory(int(sys.argv[1]))  # send end of directory to XBMC
- 
+
